+++ src/components/component/component.spec.jsx
@@ -1,6 +1,12 @@
 import React from 'react';
 import Component from './component';
-import { StyleguideDeviceFrame } from '../../../index';
+import {
+    StyleguideDeviceFrame,
+    StyleguideCell,
+    StyleguideGroup,
+    StyleguideDeviceRange,
+    StyleguidePlaceholder,
+} from '../../../index';
 
 export const SpecComponentEmpty = () => {
     return <Component name={'BasicComponent'} />;
@@ -78,5 +84,88 @@ export const SpecComponentInFrame = () => {
                 </div>
             </div>
         </StyleguideDeviceFrame>
+    );
+};
+
+export const SpecComponentInGroupCell = () => {
+    return (
+        <StyleguideGroup>
+            <StyleguideCell>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+            <StyleguideCell>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+            <StyleguideCell>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+        </StyleguideGroup>
+    );
+};
+
+export const SpecComponentInGroupCellwithLegends = () => {
+    const fontSize = 20;
+    return (
+        <StyleguideGroup>
+            <StyleguideCell legend="cell 1" fontSize={fontSize}>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+            <StyleguideCell legend="cell 2" fontSize={fontSize}>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+            <StyleguideCell legend="cell 3" fontSize={fontSize}>
+                <SpecComponentPropTypes />
+            </StyleguideCell>
+        </StyleguideGroup>
+    );
+};
+
+export const SpecComponentInGroupCellwithPlaceholder = () => {
+    const fontSize = 20;
+    return (
+        <StyleguideGroup>
+            <StyleguideCell legend="cell 1" fontSize={fontSize}>
+                <StyleguidePlaceholder
+                    placeholder="Component 1"
+                    fontSize={fontSize}
+                    backgroundColor="#ffddee"
+                />
+            </StyleguideCell>
+            <StyleguideCell legend="cell 2" fontSize={fontSize}>
+                <StyleguidePlaceholder
+                    placeholder="Component 2"
+                    fontSize={fontSize}
+                    backgroundColor="#ddeeff"
+                />
+            </StyleguideCell>
+            <StyleguideCell legend="cell 3" fontSize={fontSize}>
+                <StyleguidePlaceholder
+                    placeholder="Component 3 with default font-size"
+                    backgroundColor="#ffeedd"
+                />
+            </StyleguideCell>
+
+            <StyleguideCell legend="cell 4 with default font-size">
+                <StyleguidePlaceholder
+                    placeholder="Component 4 with default font-size"
+                    backgroundColor="#ffeedd"
+                />
+            </StyleguideCell>
+        </StyleguideGroup>
+    );
+};
+
+export const SpecComponentDeviceRange = () => {
+    const fontSize = 20;
+    return (
+        <StyleguideDeviceRange fontSize={fontSize}>
+            <StyleguideCell legend="Cloned cell across devices" fontSize={fontSize}>
+                <StyleguidePlaceholder
+                    placeholder="Cloned component across devices"
+                    fontSize={fontSize}
+                    backgroundColor="#ffddee"
+                />
+            </StyleguideCell>
+        </StyleguideDeviceRange>
     );
 };