+++ src/visual-helpers/styleguide-cell/styleguide-cell.jsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import View from '../styleguide-view';
 import Text from '../styleguide-text';
+import config from '__GLOBAL__CONFIG__';
 
 const Platform = {
     OS: 'web',
@@ -9,6 +10,7 @@ const Platform = {
 
 const propTypes = {
     legend: PropTypes.string,
+    fontSize: PropTypes.oneOf([PropTypes.number, PropTypes.string]),
     backgroundColor: PropTypes.string,
     width: PropTypes.number,
     height: PropTypes.number,
@@ -17,7 +19,15 @@ const propTypes = {
 };
 
 function StyleguideCell(props) {
-    const { legend, backgroundColor, width, height, border, children } = props;
+    const {
+        legend,
+        backgroundColor,
+        width,
+        height,
+        border,
+        fontSize = config.legendFontSize ? config.legendFontSize : 10,
+        children,
+    } = props;
 
     return (
         <View>
@@ -27,7 +37,7 @@ function StyleguideCell(props) {
                         width,
                         margin: 0,
                         marginBottom: 10,
-                        fontSize: 10,
+                        fontSize,
                         fontFamily: 'monospace',
                     }}
                 >