+++ src/visual-helpers/styleguide-placeholder/styleguide-placeholder.d.ts
@@ -2,6 +2,10 @@ import * as React from 'react';
 
 export type StyleguidePlaceholderProps = {
     /**
+     * The font-size for legend: number , px, vh or any other units
+     */
+    fontSize?: string | number;
+    /**
      * The text inside the placeholder-component
      */
     placeholder?: string;